login connected to db
