+++ apps/backend/app/api/auth.py
@@ -8,7 +8,8 @@ from fastapi import (
 )  # thingie that organizes the endpoints
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from services.auth.schemas import LoginRequest, LoginResponse
+from services.auth.login import LoginRequest, LoginResponse
+from services.auth.schemas import verify_password
 from services.auth.signup import SignupRequest, SignupResponse
 from services.database_manager.database import get_db
 from services.database_manager.managers.UserManager import user_manager
@@ -17,9 +18,13 @@ router = APIRouter(prefix='/auth', tags=['auth'])
 
 
 @router.post('/login', response_model=LoginResponse)  # so this is da login endpoint
-async def login(payload: LoginRequest):
-	return LoginResponse(message='Validation passed yaay')
-
+async def login(payload :LoginRequest, db: AsyncSession =Depends(get_db)):
+	user = await user_manager.get_by_email(db,payload.email)
+	if user is None or not verify_password(payload.password, user.hashed_password):
+		raise HTTPException(
+			status_code = status.HTTP_401_UNAUTHORIZED, detail="Invalid email or password"
+		)
+	return LoginResponse(message="Login is succesful")
 
 @router.post('/signup', response_model=SignupResponse, status_code=status.HTTP_201_CREATED)
 async def signup(request: SignupRequest, db: AsyncSession = Depends(get_db)):
