+++ services/auth/schemas.py
@@ -3,11 +3,6 @@ from __future__ import annotations
 import re
 
 import bcrypt
-from pydantic import (
-	BaseModel,
-	EmailStr,
-	field_validator,
-)  # pydantic is like a library used for validation commonly in python
 
 
 def validate_password_strength(value: str) -> str:
@@ -32,15 +27,3 @@ def verify_password(password: str, stored_hash: str) -> bool:
 	return bcrypt.checkpw(password.encode('utf-8'), stored_hash.encode('utf-8'))
 
 
-class LoginRequest(BaseModel):
-	email: EmailStr  # Pydantic library that checks if the email is the correct format..
-	password: str
-
-	@field_validator('password')
-	@classmethod
-	def validate_password(cls, value: str) -> str:
-		return validate_password_strength(value)
-
-
-class LoginResponse(BaseModel):
-	message: str
